requirements.txt
pillow
python-dotenv
azure-ai-documentintelligence
azure-ai-vision-imageanalysis
pymupdf
openai